requirements.txt
# Web framework and API
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=1.10.0

# Database
sqlalchemy>=2.0.20
pymysql>=1.1.0
cryptography>=41.0.3

# Object Storage
minio>=7.1.0

# Web scraping and HTTP
requests>=2.31.0
beautifulsoup4>=4.12.2
requests[socks]>=2.31.0
aiohttp>=3.8.5
PySocks>=1.7.1

# Image processing (with WebP support)
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Testing (development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Code quality (development)
black>=23.7.0
flake8>=6.0.0
